moves slower tests in separate file and specifies integration test phase in pom
