+++ pom.xml
@@ -17,6 +17,36 @@
         <target>1.8</target>
       </configuration>
     </plugin>
+    <plugin>
+      <groupId>org.apache.maven.plugins</groupId>
+      <artifactId>maven-surefire-plugin</artifactId>
+      <version>2.19.1</version>
+      <executions>
+        <execution>
+          <id>default-test</id>
+          <configuration>
+            <includes>
+              <include>**</include>
+            </includes>
+            <excludes>
+              <exclude>*ComputerPlayerSpeedTest*</exclude>
+            </excludes>
+          </configuration>
+        </execution>
+        <execution>
+          <id>slow-test</id>
+          <phase>integration-test</phase>
+          <goals>
+            <goal>test</goal>
+          </goals>
+          <configuration>
+            <includes>
+              <include>*ComputerPlayerSpeedTest*</include>
+            </includes>
+          </configuration>
+        </execution>
+      </executions>
+    </plugin>
   </plugins>
   </build>
   <dependencies>
